// src/app/Modules/carrera/model/cpresultatavaluacio.dto.ts
// Generated by https://quicktype.io

import { CpAvaluacioDTO } from "./cpavaluacio.dto";

export interface CpResultatAvaluacioDTO {
    adaptabilitat: number;
    ciutadania:    number;
    competencies:  number;
    compromis:     number;
    direadmin:     number;
    dirform:       number;
    dirhabilitats: number;
    dirproced:     number;
    equip:         number;
    resultats:     number;
    avaluacions:   CpAvaluacioDTO[];

}