// src/app/Modules/carrera/model/cpavaluacio.dto.ts
// Generated by https://quicktype.io

export interface CpAvaluacioDTO {
    adaptabilitat: number | null;
    ciutadania:    number | null;
    compromis:     number | null;
    direadmin:     number | null;
    dirform:       number | null;
    dirhabilitats: number | null;
    dirproced:     number | null;
    equip:         number | null;
    kavaluacio:    number;
    kparticipacio: number;
    origen:        string;
    resultats:     number | null;
    observacions:  string;
}